% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctrl_table.R
\name{ctrl_table}
\alias{ctrl_table}
\title{Create a summary table to evaluate experiment controls.}
\usage{
ctrl_table(all_data)
}
\arguments{
\item{all_data}{A merged tibble of Area, Replicate, and Treatment columns for all treatments which is made by tibbleing_loop().}
}
\value{
A summary table of experiment controls
}
\description{
Create a summary table to evaluate experiment controls.
}
