% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibbleing_loop.R
\name{tibbleing_loop}
\alias{tibbleing_loop}
\title{Create a merged tibble of Area, Replicate, and Treatment columns for all treatments.}
\usage{
tibbleing_loop(data_list, index_list, treatments)
}
\arguments{
\item{data_list}{The data list made with read_excel_file().}

\item{index_list}{A list of replicates indices that can be retrieved from experiment info list made with read_exp_info().}

\item{treatments}{A list of all treatments that can be retrieved from experiment info list made with read_exp_info().}
}
\value{
A a merged tibble of Area, Replicate, and Treatment columns for all treatments.
}
\description{
Create a merged tibble of Area, Replicate, and Treatment columns for all treatments.
}
